\name{createBufferedMatrix}
\alias{createBufferedMatrix}
\title{createBufferedMatrix}
\description{Creates a Buffered Matrix object}
\usage{createBufferedMatrix(rows, cols=0, bufferrows=1, buffercols=1,prefix="BM",directory=getwd())
}
\arguments{
  \item{rows}{Number of rows in the matrix}
  \item{cols}{Initial number of coulmns in the matrix}
  \item{bufferrows}{number of rows to be buffered if the row buffer is activated}
  \item{buffercols}{number of columns to be buffered}
  \item{prefix}{String to be used as start of name for any temporary files}
  \item{directory}{path to directory where temporary files should be stored}
}
\value{
}
\details{
}
\references{
}
\seealso{
}
\examples{
}
\author{B. M. Bolstad <bmb@bmbolstad.com>}
\keyword{}
