\name{as.BufferedMatrix}
\alias{as.BufferedMatrix}
\alias{is.BufferedMatrix}
\title{Check or Coerce object to BufferedMatrix}
\description{'as.BufferedMatrix' will coerce the supplied object into a
  BufferedMatrix. 'is.BufferedMatrix' checks whether the supplied
  argument is a BufferedMatrix.
}
\usage{
 as.BufferedMatrix(x)
 is.BufferedMatrix(x)
}
\arguments{
  \item{x}{an R object}
}
\details{
  These functions are useful for converting between R
  \code{\link{matrix}} objects and \code{\link[BufferedMatrix-class]{BufferedMatrix}} objects.
}
\author{B. M. Bolstad <bmb@bmbolstad.com>}
\keyword{manip}